Add GameBoard.vue component in frontend
A new Vue component named GameBoard.vue is added to the application frontend under src/components. This component includes features such as starting a new game, resetting the game, and joining an available game. It also displays error and general messages to the users. This adds a working game board to the application interface for the users to interact with.