Remove scheduler cleanup import from backend app
The import of the `schedule_cleanups` method from the `scheduler` module has been removed from the backend app. This may be a temporary change during testing or due to changes in the cleanup process of the appliances. This commit doesn't introduce any new functionality. The import might be reinstated in the future if necessary.